Move components to layout/views/
