--- src/app/analysis/analysis.component.html
@@ -1,1 +1,0 @@
-<p>analysis works!</p>
