+++ src/app/app.module.ts
@@ -1,15 +1,15 @@
-import { BrowserModule } from "@angular/platform-browser";
-import { NgModule } from "@angular/core";
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
 
-import { AppRoutingModule, routingComponents } from "./app-routing.module";
-import { AppComponent } from "./app.component";
+import { AppRoutingModule, routingComponents } from './app-routing.module';
+import { AppComponent } from './app.component';
 //import { DashboardComponent } from './dashboard/dashboard.component';
 //import { WizardComponent } from './wizard/wizard.component';
-import { NavbarComponent } from "./navbar/navbar.component";
-import { LoginFormComponent } from "./login-form/login-form.component";
-import { FormsModule } from "@angular/forms";
-import { HttpClientModule } from "@angular/common/http";
-import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
+import { NavbarComponent } from './layout/navbar/navbar.component';
+import { LoginFormComponent } from './login-form/login-form.component';
+import { FormsModule } from '@angular/forms';
+import { HttpClientModule } from '@angular/common/http';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';  //
 
 //Material
@@ -25,11 +25,11 @@ import {
   MatIconModule,
   MatCheckboxModule,
   MatSnackBarModule
-} from "@angular/material";
+} from '@angular/material';
 import { DashboardItemComponent } from './dashboard-item/dashboard-item.component';
-import { FooterComponent } from './footer/footer.component';
-import { HomepageComponent } from './homepage/homepage.component';
-import { AnalysisComponent } from './analysis/analysis.component';
+import { FooterComponent } from './layout/footer/footer.component';
+import { HomepageComponent } from './layout/views/homepage/homepage.component';
+import { AnalysisComponent } from './layout/views/analysis/analysis.component';
 
 
 @NgModule({
