--- src/app/dashboard/dashboard.component.html
@@ -1,20 +1,0 @@
-<div class="dashboard-component">
-  <div class="row">
-    <h2>Your Dashboard</h2>
-  </div>
-  <div class="row">
-    <div class="col-1-3"><app-dashboard-item></app-dashboard-item></div>
-    <div class="col-1-3"><app-dashboard-item></app-dashboard-item></div>
-    <div class="col-1-3"><app-dashboard-item></app-dashboard-item></div>
-  </div>
-  <div class="row">
-    <div class="col-1-3"><app-dashboard-item></app-dashboard-item></div>
-    <div class="col-1-3"><app-dashboard-item></app-dashboard-item></div>
-    <div class="col-1-3"><app-dashboard-item></app-dashboard-item></div>
-  </div>
-  <div class="row row-button">
-    <button routerLink="/wizard" class="btn-new-project" mat-flat-button>
-      Create a project
-    </button>
-  </div>
-</div>
