--- src/app/dashboard/dashboard.component.scss
@@ -1,31 +1,0 @@
-.dashboard-component {
-  background-color: #f6f6f6;
-  width: 100%;
-  height: 100%;
-  padding: 30px 30px;
-}
-
-.row {
-  width: 100%;
-}
-
-.col-1-3 {
-  width: 33.3%;
-  float: left;
-}
-
-.col-2-2 {
-  width: 70%;
-  float: left;
-}
-
-.row.row-button {
-  text-align: center;
-}
-
-.btn-new-project {
-  background-color: #ffcf6b;
-  border-radius: 50px 50px 50px 50px;
-  padding: 10px 40px;
-  margin: 20px 0;
-}
