--- src/app/homepage/homepage.component.html
@@ -1,1 +1,0 @@
-<p>homepage works!</p>
